piper_server/requirements.txt
piper-tts==1.4.1
uvicorn
fastapi[standard]
phonemizer
deep-phonemizer
dotenv
